[netstack3] Allow peek offset past ring buffer
The SendBufferWithZirconSocket reports as its readable bytes the sum of
the readable bytes in the ring buffer and the bytes in the zx socket.
That means a user of the type could reasonably try to "peek" at an
offset past the end of the ring buffer. Handle this case instead of
crashing by providing an empty slice to peek from.

Bug: 117259
Change-Id: I942725ac659565e3a01df401c48a61030e7ca66a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778742
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>