[driver_manager][DevfsExporter] Reduce repetition
Extract DevfsExporter::Export that can be shared by Export and
ExportOptions. Store exporters in a map so that they can be removed in
O(1) when devices are removed. Pass the watcher to its on_close_callback
to avoid the need to capture the pointer.

It would be even better to change `DevfsExporter::exports_` to
`std::unordered_set<ExportWatcher>` but that would require constructing
the `ExportWatcher` in-place, which would require more refactoring.

Change-Id: Ib2ba050b98ad625fe422fdff18d574cfd75693bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778962
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>