[fidl][cpp] Prepare tests for one-way calls hiding peer-closed
The ServerBindingGroup tests checks the result of a one-way call to make
sure that the remote endpoint is closed. That will soon not be the case.
Instead we pass an event handler object to the client to check it.

We can also remove the Loop::Quit calls independently of that change.
That's because Loop::RunUntilIdle would run both the client and servers
till they reach idle since they all share the same loop. Quitting the
loop early just interrupts the RunUntilIdle before things reach idle,
requiring more repeated RunUntilIdle calls.

Change-Id: I285796f82cff8d708b45f62d8aac8feb1200edb5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778283
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>