[overnet] Use a more conventional close signal in FramedStreamReader
Functions that read data from IO primitives in the Overnet code often
return a boolean indicating whether the socket closed after reading,
alongside any other returned data such as a number of bytes or decoded
packet. This contrasts with most standard IO interfaces, which return an
optional with a None value or a byte count of zero to indicate closure.

We get this habit from Quiche, which does it because it wants to expose
some information about whether the message to close a stream was in the
same packet as the last chunk of data received. However, it's an awkward
pattern in most other places, and we're better off replacing it. This
patch ends up fixing a significant chunk of PEER_CLOSED flakes that
we've been observing with CSO enabled, with the cause traced to code
insisting that certain protocol frames which should not be followed by
data are always received alongside a notice that the stream has closed.
We weren't able to keep this guarantee under CSO, and may not have been
able to keep it under packet switched Overnet either. Flakes of this
form were observed prior to CSO. They were rarer, but several
integration tests were disabled to mitigate them. A future CL will
re-enable those tests and see if the problem is gone.

Bug: 117299
Change-Id: Ie740f430b9545278abc90acd3bbdc27af0ce2a57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778331
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>