[cm] New error type: ModelError::InstanceDestroyed
Previously we used InstanceNotFound when a component
was in Destroyed state. This hides the distinction
between a component not existing in the topology and
simply being in the wrong state.

ModelError::InstanceDestroyed will now clarify this
distinction.

Change-Id: Icbfceac8d7d4ad3b16d298214dbf920d15900448
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779584
Fuchsia-Auto-Submit: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>