[netcfg] Have dhcpv6::start_client return stream
... rather than storing the stream into the StreamMap itself. Also
don't attach any context to the error on the stream to avoid double
printing it (the relevant context is readily available when an error
is yielded by the stream and already printed).

No change in behavior; this is a refactor to prepare for upcoming
changes where we also need to return a stream as a result of calling
`WatchPrefix` repeatedly.

Change-Id: I34bf0c3f6266ca2aed8ed0a70a8d642c45acb5b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762175
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Fuchsia-Auto-Submit: Tony Gong <gongt@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>