[doc-checker] Fixing link to old markdown file
The presubmit program `doc-checker` does not allow links to files in
//docs that are not at HEAD of the main branch. In this case, the
document intends to reference the older revision explicitly, so the
preferred way to implement this is to use a link to cs.opensource.google
which is allowed by doc-checker. Another alternative to consider is
to link to the Gerrit change which modified the file, if that is more
appropriate.

Bug: 117653
Change-Id: Ie0c58cd75c2dba7c8429267f13fe71b56c80803d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780091
Commit-Queue: Kayce Basques <kayce@google.com>
Fuchsia-Auto-Submit: Kayce Basques <kayce@google.com>
Reviewed-by: Kayce Basques <kayce@google.com>