[check-licenses] Fix SPDX ID creation and references.
Previously, check-licenses would try to guess what the SPDX ID for a
given license file was by using the path to that license file.

This _could_ have worked, but it was overly complicated. It was
implemented incorrectly, so the license texts were using a different ID
in the SPDX document than expected --  which causes issues when
downstream customers attempt to parse the resulting json file.

This CL updates the code to use a simple auto-incrementing number as the
IDs, and references to these IDs use the object variable directly
instead of attempting to re-generate the ID.

Tests will be added in a future CL.

Bug: 117959
Change-Id: Ib4913beb2499487f3f05b80c40d7759fec43abde
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780645
Reviewed-by: Yaar Schnitman <yaar@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>