[wlan][policy] Don't connect without interfaces.
This change modifies IfaceManager such that it will not
attempt to connect if there are no client interfaces.  The
result is that policy API clients will be notified that the
device is in an incompatible mode if they request a
connection before starting client connections.

Bug: 117485
Test: act.py \
    -c config.json \
    -tc StartStopClientConnectionsTest
Test: OOBE
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Change-Id: Ib491b764f2049652365ca2cab4dc6e6eebb90dbd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781702
Reviewed-by: Marc Khouri <mnck@google.com>
Commit-Queue: Kevin Sakuma <sakuma@google.com>