[cm] Introduce ABI reading library for in-tree Component Resolvers
Introduce ABI revision reading API for Component Resolvers to read a u64
value from a provided file path in a directory. This function does not
error if an ABI revision is not available.

This function may be used by in-tree component resolvers to read an ABI
revision and if it exists, set the value in `fresolution::Component`.

Bug: 95188
Change-Id: Ib92b462c402bda2b9ee6a226be2ee132e239430d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/754619
Commit-Queue: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>
API-Review: Gary Bressler <geb@google.com>