[build] Don't check for missing Go files in generated directories.
Failing on generated files causes build failures when a generated file
is no longer generated, as was done in fxrev.dev/779504. This forces
all Fuchsia developers to do an fx build-clean, and all bots that do
incremental builds to fail until they eventually do a clean build.

Tested this with fxrev.dev/779504 and by manually reverting this
change locally:
http://fxrev.dev/c/fuchsia/+/604261/3/tools/lib/clock/BUILD.gn
It no longer fails on the former, and continues to fail on the latter.

Fixed: 118186
Change-Id: I1eec31fcb55a8d195e68b7e1742ebfc508685fe5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781942
Commit-Queue: Cameron Dale <camrdale@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>